auth-service-django-main/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
requests==2.32.3
python-decouple==3.8
pymongo==4.6.0
PyJWT==2.8.0
gunicorn==21.2.0
django-allauth==65.0.0
django-rest-auth==0.9.5
django-email-verification==0.3.3
